refactor: アプリ本体を src/ へ分離し、APP_SCRIPT 注入ビルドと Vitest 基盤を導入
### 概要
`klondike-src.html` に内包していたアプリ本体スクリプトを `src/` 配下へ分離し、ビルド時に `{{APP_SCRIPT}}` として注入する構成へ移行しました。
合わせて `Rules` / `Solver` の分割、`Deal Again` の挙動統一、Vitest 導入と最小ユニットテスト追加を実施しています。

### 変更内容
- ビルド構成の拡張
  - `scripts/build.mjs` に `{{APP_SCRIPT}}` トークン対応を追加
  - `src/main.ts` + `src/core/solver.ts` + `src/core/rules.ts` を連結して HTML に注入
- スクリプト分離
  - `klondike-src.html` から本体 JS を除去し、`{{APP_SCRIPT}}` を配置
  - 新規: `src/main.ts`
  - 新規: `src/core/solver.ts`
  - 新規: `src/core/rules.ts`
- ゲーム挙動の調整
  - `Deal Again` を `location.reload()` から `Controller.startNewGame()` 呼び出しへ変更
  - 保存タイミングを調整（手動操作直後の中間保存を避け、自動連鎖終了時に保存）
- テスト基盤追加
  - `vitest`, `jsdom` を devDependencies に追加
  - `package.json` に `test:unit` / `test:unit:watch` を追加
  - 新規: `vitest.config.ts`
  - 新規テスト:
    - `test/rules.spec.ts`
    - `test/solver.spec.ts`
    - `test/helpers/load-core.ts`
- ドキュメント更新
  - `README.md`: `src/main.ts` をディレクトリ構成へ追記
  - `ARCHITECTURE.md`: 再開メモおよび進捗 TODO を更新

### 変更ファイル
- `ARCHITECTURE.md`
- `README.md`
- `klondike-src.html`
- `klondike.html`
- `klondike-online.html`
- `scripts/build.mjs`
- `package.json`
- `package-lock.json`
- `src/main.ts` (new)
- `src/core/solver.ts` (new)
- `src/core/rules.ts` (new)
- `vitest.config.ts` (new)
- `test/helpers/load-core.ts` (new)
- `test/rules.spec.ts` (new)
- `test/solver.spec.ts` (new)

### 確認項目
- `npm run build` が成功すること
- `npm run test:unit` が成功すること
- `klondike.html` / `klondike-online.html` の起動が継続して可能なこと
- `Deal Again` が `New Game` 相当の挙動になること
