+++ src/main.ts
@@ -1,149 +1,3 @@
-<!doctype html>
-<html lang="ja">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no, viewport-fit=cover" />
-    <title>Klondike (Solitaire)</title>
-    <script src="https://cardmeister.github.io/elements.cardmeister.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/motion@11.11.13/dist/motion.min.js"></script>
-    <style>
-      :root {
-        --card-w: 100px; --card-h: 140px;
-        --board-pad: 24px;
-        --grid-gap: 15px;
-        --bg: #008000;
-        --accent: #ffffff;
-        --ink: #f5f2ea;
-      }
-      * { box-sizing: border-box; }
-      body {
-        margin: 0; font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif; color: var(--ink);
-        background: var(--bg);
-        min-height: 100svh; overflow: hidden;
-        padding-top: env(safe-area-inset-top);
-        user-select: none; -webkit-user-select: none;
-        touch-action: manipulation;
-      }
-      @supports (height: 100dvh) {
-        body { min-height: 100dvh; }
-      }
-      header { padding: 15px 24px; display: flex; align-items: center; justify-content: center; flex-wrap: wrap; gap: 10px; max-width: 900px; margin: 0 auto; }
-      h1 { margin: 0; font-size: 18px; color: var(--accent); font-weight: normal; text-shadow: 1px 1px 2px black; text-align: center; display: flex; align-items: center; justify-content: center; gap: 6px; }
-      .app-version { font-size: 12px; opacity: 0.8; }
-      .github-link { display: inline-flex; align-items: center; color: var(--accent); opacity: 0.9; }
-      .github-link:hover { opacity: 1; }
-      .github-link svg { width: 16px; height: 16px; display: block; }
-      .controls { display: flex; gap: 12px; align-items: center; flex-wrap: wrap; justify-content: center; }
-      
-      .toggle-label { display: flex; align-items: center; cursor: pointer; color: white; font-size: 12px; gap: 5px; background: rgba(0,0,0,0.2); padding: 4px 8px; border-radius: 4px; }
-      .toggle-label input { cursor: pointer; }
-
-      button { background: #ece9d8; color: black; border: 1px solid #003c74; padding: 6px 12px; border-radius: 3px; font-size: 12px; cursor: pointer; white-space: nowrap; }
-      button:hover { box-shadow: 0 0 0 1px #f5f2ea inset; }
-      button:active { transform: translateY(1px); }
-      button:disabled { opacity: 0.5; cursor: not-allowed; }
-      
-      #status-msg { color: #ffff00; font-weight: bold; font-size: 14px; margin-right: 10px; text-shadow: 1px 1px 0 #000; min-width: 80px; text-align: right; }
-      
-      .board { padding: 10px var(--board-pad) 40px; display: grid; gap: 40px; max-width: 900px; margin: 0 auto; transition: gap 0.3s; }
-      .legal { text-align: center; font-size: 12px; margin: 12px 0 16px; opacity: 0.85; border-top: 1px solid rgba(255,255,255,0.25); padding-top: 8px; display: flex; align-items: center; justify-content: center; gap: 8px; flex-wrap: nowrap; }
-      .legal a { color: var(--accent); text-decoration: underline; margin: 0 6px; vertical-align: middle; }
-      .legal a.oss-badge { text-decoration: none; }
-      .oss-badge { display: inline-flex; align-items: center; justify-content: center; padding: 2px 8px; border-radius: 999px; background: rgba(0,0,0,0.25); border: 1px solid rgba(255,255,255,0.35); font-size: 10px; font-weight: bold; color: var(--accent); margin: 0 6px; white-space: nowrap; }
-      .top-row, .bottom-row { display: grid; grid-template-columns: repeat(7, var(--card-w)); gap: var(--grid-gap); justify-content: center; transition: gap 0.3s; }
-
-      @media (max-width: 800px) {
-        :root {
-          --board-pad: 10px;
-          --grid-gap: 4px;
-          --card-w: clamp(34px, calc((100vw - (var(--board-pad) * 2) - (var(--grid-gap) * 6)) / 7), 100px);
-          --card-h: calc(var(--card-w) * 1.4);
-        }
-        header { padding: 12px 10px; }
-        .controls { width: 100%; }
-        .board { padding: 10px var(--board-pad); gap: 20px; }
-      }
-
-      .slot, .column {
-        position: relative; width: var(--card-w); min-height: var(--card-h);
-        border-radius: 4px; 
-        border: 2px solid rgba(255,255,255,0.2);
-        cursor: pointer; 
-        transition: height 0.2s;
-      }
-      .slot:empty, .column:empty { border: 2px solid rgba(0,0,0,0.1); background: rgba(0,0,0,0.05); }
-
-      playing-card { position: absolute; width: var(--card-w); height: var(--card-h); cursor: pointer; filter: drop-shadow(1px 1px 3px rgba(0,0,0,0.6)); transition: top 0.1s; pointer-events: auto; }
-      .stock-count { display:none; }
-      
-      #win-overlay {
-        display: none; position: fixed; inset: 0; background: rgba(0,0,0,0.6);
-        z-index: 100; flex-direction: column; align-items: center; justify-content: center;
-        color: white; animation: fadeIn 0.5s forwards;
-      }
-      @keyframes fadeIn { from { opacity: 0; } to { opacity: 1; } }
-      .win-msg { font-size: 40px; margin-bottom: 20px; font-family: "Times New Roman", serif; }
-      
-      #stuck-warning {
-        position: fixed; top: 15%; left: 50%; transform: translateX(-50%);
-        background: rgba(255, 0, 0, 0.85); color: white; padding: 10px 20px;
-        border-radius: 20px; font-weight: bold; pointer-events: none;
-        opacity: 0; transition: opacity 0.3s; z-index: 50;
-        box-shadow: 0 4px 10px rgba(0,0,0,0.5); font-size: 14px;
-        white-space: nowrap;
-      }
-    </style>
-  </head>
-  <body oncontextmenu="return false;">
-    <div id="win-overlay">
-        <div class="win-msg">Congratulations</div>
-        <button id="btn-deal-again">Deal Again</button>
-    </div>
-    
-    <div id="stuck-warning">⚠️ Likely Stuck (Undo Recommended)</div>
-
-    <header>
-      <h1>
-        Klondike (Solitaire)
-        <span id="app-version" class="app-version"></span>
-      </h1>
-      <div class="controls">
-        <span id="status-msg"></span>
-        <label class="toggle-label" title="Auto check + debug logs">
-            <input type="checkbox" id="chk-autocheck" checked> Solvability Check
-        </label>
-        <button id="btn-undo" disabled>Undo</button>
-        <button id="btn-restart" disabled>Restart</button>
-        <button id="btn-hint">Hint</button>
-        <button id="new-game">New Game</button>
-        <button id="btn-test">Test</button>
-      </div>
-    </header>
-    <main class="board">
-      <section class="top-row">
-        <div id="stock" class="slot"></div><div id="waste" class="slot"></div>
-        <div class="spacer"></div>
-        <div id="foundation-0" class="slot"></div><div id="foundation-1" class="slot"></div>
-        <div id="foundation-2" class="slot"></div><div id="foundation-3" class="slot"></div>
-      </section>
-      <section class="bottom-row">
-        <div id="tableau-0" class="column"></div><div id="tableau-1" class="column"></div>
-        <div id="tableau-2" class="column"></div><div id="tableau-3" class="column"></div>
-        <div id="tableau-4" class="column"></div><div id="tableau-5" class="column"></div>
-        <div id="tableau-6" class="column"></div>
-      </section>
-    </main>
-    <footer class="legal">
-      <span>Copyright (c) Toshiki Iga</span>
-      <a class="github-link" href="https://github.com/igapyon/klondike/" aria-label="GitHub Repository" target="_blank" rel="noopener">
-        <svg viewBox="0 0 16 16" aria-hidden="true">
-          <path fill="currentColor" d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27s1.36.09 2 .27c1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.01 8.01 0 0 0 16 8c0-4.42-3.58-8-8-8Z"/>
-        </svg>
-      </a>
-      <a class="oss-badge" href="https://github.com/igapyon/klondike/blob/main/THIRD_PARTY_NOTICES.md" target="_blank" rel="noopener" aria-label="Third-party notices">Third-Party Notices</a>
-    </footer>
-
-        <script>
       /* --- [Config] --- */
       const Config = {
         APP_VERSION: "v20260304",
@@ -1073,223 +927,3 @@
           }
         });
       });
-
-
-/* --- [Simulator] --- */
-      class Simulator {
-        constructor() {}
-        clone(s) { return JSON.parse(JSON.stringify(s)); }
-        
-        findSolvableDeck() {
-            let attempts = 0;
-            while (attempts < 500) {
-                const deck = this.createDeck();
-                const startState = this.deal(deck);
-                if (this.canSolve(this.clone(startState), { limit: Config.SOLVER_LIMIT_DEAL, noProgressLimit: Config.SOLVER_STALL_DEAL })) return startState;
-                attempts++;
-            }
-            return this.deal(this.createDeck());
-        }
-
-        checkCurrentState(currentState) {
-            for(let i=0; i<13; i++) {
-                if (this.canSolve(this.clone(currentState), { limit: Config.SOLVER_LIMIT_CHECK, noProgressLimit: Config.SOLVER_STALL_CHECK })) return true;
-            }
-            return false;
-        }
-
-        createDeck() {
-            const d = [];
-            // 【変更点2】 色の判定ロジックを変更。偶数(0,2)が黒、奇数(1,3)が赤
-            for (let s = 0; s < 4; s++) for (let r = 0; r < 13; r++) d.push({ suit: s, rank: r + 1, color: (s % 2 === 0) ? 0 : 1, isOpen: false });
-            for (let i = d.length - 1; i > 0; i--) { const j = Math.floor(Math.random() * (i + 1)); [d[i], d[j]] = [d[j], d[i]]; }
-            return d;
-        }
-        deal(d) {
-             let t = Array.from({ length: 7 }, () => []);
-             const tempDeck = [...d];
-             for (let i = 0; i < 7; i++) {
-                 for (let j = 0; j <= i; j++) {
-                     const c = tempDeck.pop(); c.isOpen = (j === i); t[i].push(c);
-                 }
-             }
-             return { stock: tempDeck, waste: [], foundations: [[],[],[],[]], tableau: t };
-        }
-
-        canSolve(simState, opts = {}) {
-            let moves = 0;
-            let noProgressLoop = 0;
-            const limit = (opts.limit ?? Config.SOLVER_LIMIT_CHECK);
-            const noProgressLimit = (opts.noProgressLimit ?? Config.SOLVER_STALL_CHECK);
-            while (moves < limit && noProgressLoop < noProgressLimit) {
-                let moved = false;
-                for (let i = 0; i < 7; i++) {
-                    const col = simState.tableau[i];
-                    if (col.length > 0) {
-                        const c = col[col.length - 1];
-                        if (this.checkF(simState, c, c.suit)) {
-                            simState.foundations[c.suit].push(col.pop());
-                            if (col.length > 0) col[col.length - 1].isOpen = true;
-                            moved = true; break;
-                        }
-                    }
-                }
-                if (moved) { moves++; noProgressLoop=0; continue; }
-                if (simState.waste.length > 0) {
-                    const c = simState.waste[simState.waste.length - 1];
-                    if (this.checkF(simState, c, c.suit)) {
-                        simState.foundations[c.suit].push(simState.waste.pop());
-                        moved = true;
-                    }
-                }
-                if (moved) { moves++; noProgressLoop=0; continue; }
-                for (let i = 0; i < 7; i++) {
-                    const col = simState.tableau[i];
-                    if (col.length === 0) continue;
-                    let openIdx = -1;
-                    for(let k=0; k<col.length; k++) { if(col[k].isOpen) { openIdx=k; break; } }
-                    if (openIdx === -1) continue; 
-                    if (openIdx === 0 && col[openIdx].rank === 13) continue;
-                    const cardToMove = col[openIdx];
-                    const target = this.findTabTarget(simState, cardToMove, i);
-                    if (target !== -1) {
-                        const run = col.splice(openIdx);
-                        simState.tableau[target] = simState.tableau[target].concat(run);
-                        if (col.length > 0) col[col.length - 1].isOpen = true;
-                        moved = true; break;
-                    }
-                }
-                if (moved) { moves++; noProgressLoop=0; continue; }
-                if (simState.waste.length > 0) {
-                    const c = simState.waste[simState.waste.length - 1];
-                    const target = this.findTabTarget(simState, c, -1);
-                    if (target !== -1) {
-                        simState.tableau[target].push(simState.waste.pop());
-                        moved = true;
-                    }
-                }
-                if (moved) { moves++; noProgressLoop=0; continue; }
-                if (simState.stock.length > 0) {
-                    const c = simState.stock.pop(); c.isOpen = true;
-                    simState.waste.push(c); moved = true;
-                } else if (simState.stock.length === 0 && simState.waste.length > 0) {
-                    simState.stock = simState.waste.reverse();
-                    simState.stock.forEach(x => x.isOpen = false);
-                    simState.waste = []; moved = true;
-                }
-                if (moved) {
-                     moves++; 
-                     if (simState.stock.length === 0 && simState.waste.length === 0) noProgressLoop = 0;
-                     else noProgressLoop++;
-                } else break;
-            }
-            return simState.foundations.reduce((a, b) => a + b.length, 0) === 52;
-        }
-        getNextMove(simState) {
-            for (let i = 0; i < 7; i++) {
-                const col = simState.tableau[i];
-                if (col.length > 0) {
-                    const c = col[col.length - 1];
-                    if (this.checkF(simState, c, c.suit)) {
-                        return { type: "t2f", tableauIdx: i, suit: c.suit };
-                    }
-                }
-            }
-            if (simState.waste.length > 0) {
-                const c = simState.waste[simState.waste.length - 1];
-                if (this.checkF(simState, c, c.suit)) {
-                    return { type: "w2f", suit: c.suit };
-                }
-            }
-            for (let i = 0; i < 7; i++) {
-                const col = simState.tableau[i];
-                if (col.length === 0) continue;
-                let openIdx = -1;
-                for (let k = 0; k < col.length; k++) { if (col[k].isOpen) { openIdx = k; break; } }
-                if (openIdx === -1) continue;
-                if (openIdx === 0 && col[openIdx].rank === 13) continue;
-                const cardToMove = col[openIdx];
-                const target = this.findTabTarget(simState, cardToMove, i);
-                if (target !== -1) {
-                    return { type: "t2t", fromIdx: i, toIdx: target, openIdx };
-                }
-            }
-            if (simState.waste.length > 0) {
-                const c = simState.waste[simState.waste.length - 1];
-                const target = this.findTabTarget(simState, c, -1);
-                if (target !== -1) {
-                    return { type: "w2t", toIdx: target };
-                }
-            }
-            if (simState.stock.length > 0) return { type: "draw" };
-            if (simState.stock.length === 0 && simState.waste.length > 0) return { type: "recycle" };
-            return null;
-        }
-        checkF(s, card, fIdx) {
-            const f = s.foundations[fIdx];
-            if (f.length === 0) return card.rank === 1 && card.suit === fIdx;
-            const top = f[f.length - 1];
-            return card.suit === top.suit && card.rank === top.rank + 1;
-        }
-        findTabTarget(s, card, ignoreIdx) {
-            for (let i = 0; i < 7; i++) {
-                if (i === ignoreIdx) continue;
-                const dest = s.tableau[i];
-                if (dest.length === 0) { if (card.rank === 13) return i; } 
-                else {
-                    const top = dest[dest.length - 1];
-                    if (card.rank === top.rank - 1 && card.color !== top.color) return i;
-                }
-            }
-            return -1;
-        }
-      }
-      const bot = new Simulator();
-
-      
-
-/* --- [Rules] --- */
-      function isSafeToAutoMove(card) {
-        const state = State.current;
-        if (card.rank === 1) return true; 
-        const targetRank = card.rank - 1;
-        const otherColor = 1 - card.color;
-        let count = 0;
-        state.foundations.forEach((f, idx) => {
-            // 【変更点3】 色判定ロジックの修正 (0=S黒, 1=H赤, 2=C黒, 3=D赤)
-            const fColor = (idx % 2 === 0) ? 0 : 1;
-            if (fColor === otherColor) {
-                const top = f[f.length - 1];
-                if (top && top.rank >= targetRank) count++;
-            }
-        });
-        return count === 2; 
-      }
-      function getCid(card) {
-        if (!card || !card.isOpen) return "back"; 
-        return `${Config.RANKS[card.rank - 1]}-of-${Config.SUITS[card.suit]}`;
-      }
-      function getCardKey(card) { return `${card.suit}-${card.rank}`; }
-      function snapshotCard(card) {
-        if (!card) return null;
-        return { suit: card.suit, rank: card.rank, color: card.color, isOpen: card.isOpen };
-      }
-      function checkFoundationMove(card, fIdx) {
-        const found = State.current.foundations[fIdx];
-        if (found.length === 0) return { ok: (card.rank === 1 && card.suit === fIdx) };
-        const top = found[found.length - 1];
-        return { ok: (card.suit === top.suit && card.rank === top.rank + 1) };
-      }
-      function canPlaceOnTableau(card, tIdx) {
-        const dest = State.current.tableau[tIdx];
-        if (dest.length === 0) return { ok: (card.rank === 13), type: "K-only" };
-        const top = dest[dest.length - 1];
-        const match = top.isOpen && (card.rank === top.rank - 1) && (card.color !== top.color);
-        return match ? { ok: true, type: "Run" } : { ok: false };
-      }
-      const Rules = { isSafeToAutoMove, getCid, checkFoundationMove, canPlaceOnTableau };
-
-      
-</script>
-  </body>
-</html>
